+++ tests/api/test_api_modules.py
@@ -1,6 +1,7 @@
 import json
 import unittest
 import uuid
+import warnings
 
 from requests import Response, Session
 from unittest.mock import patch
@@ -145,6 +146,55 @@ class TestAPIPersonsModule(TestAPIModule):
         self.assertEqual(response.json(), resp_body)
         self.assertEqual(response.status_code, status_code)
 
+    @patch.object(Session, 'send')
+    def test_persons_entry(self, mocked_send):
+        status_code = 201
+        resp_body = {'idxid': str(uuid.uuid4())}
+        mocked_send.return_value = self.response(resp_body, status_code)
+        response = self.client.persons.entry(id=1)
+        self.assertEqual(response.json(), resp_body)
+        self.assertEqual(response.status_code, status_code)
+
+
+class TestAPIEntriesModule(TestAPIModule):
+
+    @patch.object(Session, 'send')
+    def test_list_entries(self, mocked_send):
+        status_code = 200
+        resp_body = {'key': 'value'}
+        mocked_send.return_value = self.response(resp_body, status_code)
+        response = self.client.entries.list(
+            conf='new,exact,ha', limit=10)
+        self.assertEqual(response.json(), resp_body)
+        self.assertEqual(response.status_code, status_code)
+
+    @patch.object(Session, 'send')
+    def test_stats_idxid(self, mocked_send):
+        status_code = 200
+        resp_body = {'key': 'value'}
+        mocked_send.return_value = self.response(resp_body, status_code)
+        response = self.client.entries.stats_idxid(idxid='1')
+        self.assertEqual(response.json(), resp_body)
+        self.assertEqual(response.status_code, status_code)
+
+    @patch.object(Session, 'send')
+    def test_stats_sources(self, mocked_send):
+        status_code = 200
+        resp_body = {'key': 'value'}
+        mocked_send.return_value = self.response(resp_body, status_code)
+        response = self.client.entries.stats_sources(conf='new', limit=1000)
+        self.assertEqual(response.json(), resp_body)
+        self.assertEqual(response.status_code, status_code)
+
+    @patch.object(Session, 'send')
+    def test_delete_entry(self, mocked_send):
+        status_code = 204
+        resp_body = None
+        mocked_send.return_value = self.response(resp_body, status_code)
+        response = self.client.entries.delete(id=1)
+        self.assertEqual(response.json(), resp_body)
+        self.assertEqual(response.status_code, status_code)
+
 
 class TestAPIRecordsModule(TestAPIModule):
 
@@ -153,8 +203,15 @@ class TestAPIRecordsModule(TestAPIModule):
         status_code = 200
         resp_body = {'key': 'value'}
         mocked_send.return_value = self.response(resp_body, status_code)
-        response = self.client.records.list(
-            new=True, junk=False, ha=False, qty=10)
+
+        with warnings.catch_warnings(record=True) as w:
+            response = self.client.records.list(
+                new=True, junk=False, ha=False, qty=10)
+
+            assert len(w) == 1
+            assert issubclass(w[-1].category, FutureWarning)
+            assert 'deprecated' in str(w[-1].message).lower()
+
         self.assertEqual(response.json(), resp_body)
         self.assertEqual(response.status_code, status_code)
 
@@ -163,7 +220,14 @@ class TestAPIRecordsModule(TestAPIModule):
         status_code = 200
         resp_body = {'key': 'value'}
         mocked_send.return_value = self.response(resp_body, status_code)
-        response = self.client.records.get(idxid='1')
+
+        with warnings.catch_warnings(record=True) as w:
+            response = self.client.records.get(idxid='1')
+
+            assert len(w) == 1
+            assert issubclass(w[-1].category, FutureWarning)
+            assert 'deprecated' in str(w[-1].message).lower()
+
         self.assertEqual(response.json(), resp_body)
         self.assertEqual(response.status_code, status_code)
 
@@ -172,7 +236,14 @@ class TestAPIRecordsModule(TestAPIModule):
         status_code = 200
         resp_body = {'key': 'value'}
         mocked_send.return_value = self.response(resp_body, status_code)
-        response = self.client.records.entry_delete(entry_id=1)
+
+        with warnings.catch_warnings(record=True) as w:
+            response = self.client.records.entry_delete(entry_id=1)
+
+            assert len(w) == 1
+            assert issubclass(w[-1].category, FutureWarning)
+            assert 'deprecated' in str(w[-1].message).lower()
+
         self.assertEqual(response.json(), resp_body)
         self.assertEqual(response.status_code, status_code)